Struct simplot::Figure

pub struct Figure {
    // some fields omitted
}

Plot container

Methods

impl Figure

fn new() -> Figure

Creates an empty figure

fn draw(&mut self) -> Result<Child>

Spawns a drawing child process

NOTE: stderr, stdin, and stdout are piped

fn dump<W>(&mut self, sink: &mut W) -> Result<&mut Figure> where W: Write

Dumps the script required to produce the figure into sink

fn save(&mut self, path: &Path) -> Result<&mut Figure>

Saves the script required to produce the figure to path